Run Importer: Retry Mark Fail when Import Program error exit
When Import Program exit with error code but import tracking record is
still "running", the Run Import program marks the import tracking record
as failed and the import tracking run record as failed.
Now if it gets an error updating the DB for setting those records to
failed, it will continue to retry until it succeeds.
This will handle DB restarts.  This will not handle server restarts
where the DB is stopped before the run import program.