Add Isotope Labeling for Peptides and Proteins, Phase 1
This is Phase 1.
The isotope label is accepted in the Proxl XML file and stored in the
database.
The isotope label is used for Protein inference.  If the peptide is
labeled, it only matches proteins with the same label.  If the peptide
is not labeled, it only matches proteins with no label.

A peptide is part of a reported peptides.

This implements the proxl-import-api V1.6.0.

The web app and importer must be down while the SQL upgrade is applied.
The previous web app will not work after the SQL upgrade is applied.