QC Charts drawn while section is closed fix (many fixed)
If a section is closed while a QC chart is drawn, it is drawn the wrong
size since using the containing div size to control the chart.
Hold on to the data and draw the chart the next time the section is
opened.
Complete for all of Merged, and on Single Search the Summary Statistic
section.