Update DB Install Script with Comments for proper regen from Workbench
There is a change to table unified_rep_pep_dynamic_mod_lookup, field
mass_rounded_string to remove the COLLATE but that was not required
since the table has the same COLLATE.
This is likely an artifact of the change of MySQL Workbench Program
version.
The field has never had a COLLATE specified in the MySQL Workbench file.
The COLLATE was added by the MySQL Workbench Program when new fields
were added to the table.